+++ icu4c/source/test/intltest/idnaconf.cpp
@@ -73,7 +73,7 @@ UBool IdnaConfTest::ReadAndConvertFile(){
     }
 
     const char* name = "idna_conf.txt";     // test data file
-    int t = strlen(path) + strlen(name) + 1;
+    int t = static_cast<int>(strlen(path) + strlen(name) + 1);
     char* absolute_name = new char[t];
     strcpy(absolute_name, path);
     strcat(absolute_name, name);
@@ -108,14 +108,14 @@ UBool IdnaConfTest::ReadAndConvertFile(){
                                 NULL,           //  dest,
                                 0,              //  destCapacity,
                                 source,
-                                source_len,
+                                static_cast<int32_t>(source_len),
                                 &status);
     if (status == U_BUFFER_OVERFLOW_ERROR) {
         // Buffer Overflow is expected from the preflight operation.
         status = U_ZERO_ERROR;
         UChar * dest = NULL;
         dest = new UChar[ dest_len + 1];
-        ucnv_toUChars(conv, dest, dest_len + 1, source, source_len, &status);
+        ucnv_toUChars(conv, dest, dest_len + 1, source, static_cast<int32_t>(source_len), &status);
         // Do not know the "if possible" behavior of ucnv_toUChars()
         // Do it by ourself.
         dest[dest_len] = 0;
